+++ src/server/game/Spells/Auras/SpellAuras.cpp
@@ -2079,8 +2079,8 @@ void Aura::HandleAuraSpecificMods(AuraApplication const* aurApp, Unit* caster, b
                     {
                         uint32 damage = caster->SpellDamageBonusDone(target, GetSpellInfo(), GetEffect(0)->GetAmount(), DOT, EFFECT_0);
                         damage = target->SpellDamageBonusTaken(caster, GetSpellInfo(), damage);
-                        float basepoints0 = aurEff->GetAmount() * GetEffect(0)->GetTotalTicks() * int32(damage) / 100;
-                        float heal = int32(CalculatePct(basepoints0, 15));
+                        float basepoints0 = aurEff->GetAmount() * GetEffect(0)->GetTotalTicks() * damage / 100.0f;
+                        float heal = CalculatePct(basepoints0, 15.0f);
 
                         caster->CastCustomSpell(target, 63675, &basepoints0, NULL, NULL, true, NULL, GetEffect(0));
                         caster->CastCustomSpell(caster, 75999, &heal, NULL, NULL, true, NULL, GetEffect(0));
